pyproject.toml
[tool.poetry]
name = "cloudsend"
version = "0.1.0"
description = "CloudSend is a Python package that allows you to run any script on a cloud service (for now AWS only)."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/benbenz/cloudsend"

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1"
jcs = "^0.2.1"
pyyaml = "^6.0"
paramiko = "^2.12.0"
ortools = "^9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
demo = 'cloudsend.demo:main'
cli = 'cloudsend.cli:main'

[tool.pytest.ini_options]
asyncio_mode="strict"
testpaths = [
    "tests"
]
